Further documentation and input changed
Refactored the classes further and documented
the classes.

Updated the input for the neural network.
It now sends the x_end and y_midpoint of
the upcoming obstacle and the distance between
that at point and second upcoming obstacle's
x_start and y_midpoint

Also sends y co-ord of bird like before.
Changed dir for weights.