Added supervised model to train the bird
Adding supervised trainer where instead of genetic algorithm
it will do logistic regression ttrain a model and use
that to deicde whether to jump or not.
Lets the user play at first to get initial data to train
the model oon, and then on every round adds more data.
The model is still very poor and requires further debuggin
to make it improve each round.